New: Add ErrorPacket for reading errors from the discord client
This commit adds a new packet, opcode 2 which is received when the discord client ("server") has an error with one of our packets. This could be due to an invalid rich presence object, application id, or something else. Instead of treating errors as unknown packets, let's treat them as they should be.