Rewrite the packet encoding / decoding pipeline
The old pipeline was quite hard to understand, with some packets being mistreated.

Now, we have a better system where packets do no basically no encoding / decoding. All the encoding or decoding is handled by a PacketHandler.
This system is a little patched together (due to weird generics stuff), however it is better than our previous system. To add a packet handler, you add to the handlers set in PacketFactory.

We need to handle event type packets correctly, as right now we are not doing that.